% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{mlr3learners.plugdensity-package}
\alias{mlr3learners.plugdensity}
\alias{mlr3learners.plugdensity-package}
\title{mlr3learners.plugdensity: Learners from the {plugdensity} package for 'mlr3'}
\description{
Adds learners from the {plugdensity} package to
    'mlr3'.
}
\author{
\strong{Maintainer}: Raphael Sonabend \email{raphael.sonabend.15@ucl.ac.uk}

}
